enh: 拆开逻辑
